refactor(#106): Move file cache implementation to separate file
