+++ internal/cache/aidy_cache_test.go
